created separate namespaces for decorators and definitions
