+++ python/lsst/sims/catalogs/__init__.py
@@ -1,6 +1,2 @@
 import pkgutil
 __path__ = pkgutil.extend_path(__path__,__name__)
-
-from .decorators import *
-from .InstanceCatalog import *
-from .CompoundInstanceCatalog import *
