+++ python/lsst/sims/catalogs/db/dbConnection.py
@@ -569,7 +569,7 @@ class CatalogDBObject(DBObject):
 
     def getCatalog(self, ftype, *args, **kwargs):
         try:
-            from lsst.sims.catalogs import InstanceCatalog
+            from lsst.sims.catalogs.definitions import InstanceCatalog
             return InstanceCatalog.new_catalog(ftype, self, *args, **kwargs)
         except ImportError:
             raise ImportError("sims_catalogs not set up.  Cannot get InstanceCatalog from the object.")
