+++ python/lsst/sims/catalogs/decorators/__init__.py
@@ -1,0 +1,1 @@
+from .decorators import *
