+++ python/lsst/sims/catalogs/definitions/CompoundInstanceCatalog.py
