+++ python/lsst/sims/catalogs/definitions/InstanceCatalog.py
