+++ python/lsst/sims/catalogs/definitions/__init__.py
@@ -1,0 +1,2 @@
+from .InstanceCatalog import *
+from .CompoundInstanceCatalog import *
